pyproject.toml
[project]
name = "dns-collector"
version = "0.1.0"
description = "DNS records ingestion pipeline"
requires-python = ">=3.13"
dependencies = [
    "dnspython>=2.6,<3.0",
    "duckdb>=1.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff>=0.4",
    "ty",
    "interrogate>=1.7",
    "pre-commit>=3.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.interrogate]
ignore-init-method = true
ignore-magic = true
fail-under = 80
verbose = 1

[tool.pytest.ini_options]
testpaths = ["tests"]
